* implement Resource.from_dict() (WIP)
