+++ notion_df/resource/schema.py
@@ -1,6 +1,6 @@
 from abc import ABCMeta
 
-from notion_df.resource import Resource
+from notion_df.resource.resource import Resource
 
 
 class PropertySchema(Resource, metaclass=ABCMeta):
